Criando service para salvar dados de satelites no banco de dados
