+++ src/main/java/com/projeto/satellite/GeostationarySatellite/Exceptions/ValidatingDuplicateValues.java
@@ -1,0 +1,9 @@
+package com.projeto.satellite.GeostationarySatellite.Exceptions;
+
+public class ValidatingDuplicateValues extends RuntimeException {
+	private static final long serialVersionUID = 1L;
+
+	public ValidatingDuplicateValues(String msg) {
+		super(String.format(msg));
+ 	}
+}
